Fixed issue with cervical rhythm link not beeing included in Rigid thoracic model.
The cervical link driver now connects the cervical rhythm to the entire spine motion. This way it will also work even though the rigid thorax is used.

This commit also modfies the links rhythm coefficients so they are no longer recalculated as as ratios, but instead uses the values of the coefficient from excel sheet directly. This makes it easier for us to see where the values come from