Reworked the shoulder rhythm It now ignores negative elevation planes there was no data for these
A starting guess on the ThoraxHumerus rotational measure that has a positive value for elevation has been added, this makes is more likely that among the rotational solutions it will pick the solution with a positive elevation.

Additionally it the starting guess is not enough to ensure positive elevation the ShoulderRhythm  is using a "processed" version of the thoraxHumerus measure.
This processed version flips all signs if elevation plane is negative

The rhythm code is also restructured to be wrapped into AnyKinEq